+++ examples/lifetime_ext.rs
@@ -3,6 +3,9 @@ use traitful::extend;
 pub trait Lifetime<'a> {}
 
 #[extend(for<T: Lifetime<'a>> T)]
-pub trait MaybeGeneric<'a, U>: Lifetime<'a> {}
+pub trait ExtensionA<'a, U>: Lifetime<'a> {}
+
+#[extend]
+pub trait ExtensionB<'b>: Lifetime<'b> {}
 
 fn main() {}