+++ src/common/obfuscate_generics.rs
@@ -1,6 +1,6 @@
 use syn::{
-    punctuated::Punctuated, token::Comma, CapturedParam, GenericParam,
-    Generics, Ident, TypeParamBound,
+    CapturedParam, GenericParam, Generics, Ident, TypeParamBound,
+    punctuated::Punctuated, token::Comma,
 };
 
 /// Obfuscate user-defined generics, so they don't conflict with traitful's
@@ -14,29 +14,29 @@ pub(crate) fn obfuscate_generics(mut generics: Generics) -> Generics {
 fn obfuscate_params(params: &mut Punctuated<GenericParam, Comma>) {
     for param in params.iter_mut() {
         match param {
-            GenericParam::Lifetime(ref mut param) => {
+            GenericParam::Lifetime(param) => {
                 obfuscate_ident(&mut param.lifetime.ident);
             }
-            GenericParam::Type(ref mut param) => {
+            GenericParam::Type(param) => {
                 obfuscate_ident(&mut param.ident);
 
                 for bound in param.bounds.iter_mut() {
                     match bound {
-                        TypeParamBound::Trait(ref mut bound) => {
+                        TypeParamBound::Trait(bound) => {
                             if let Some(ref mut lifetimes) = bound.lifetimes {
                                 obfuscate_params(&mut lifetimes.lifetimes);
                             }
                         }
-                        TypeParamBound::Lifetime(ref mut bound) => {
+                        TypeParamBound::Lifetime(bound) => {
                             obfuscate_ident(&mut bound.ident);
                         }
-                        TypeParamBound::PreciseCapture(ref mut bound) => {
+                        TypeParamBound::PreciseCapture(bound) => {
                             for param in bound.params.iter_mut() {
                                 match param {
-                                    CapturedParam::Lifetime(ref mut param) => {
+                                    CapturedParam::Lifetime(param) => {
                                         obfuscate_ident(&mut param.ident);
                                     }
-                                    CapturedParam::Ident(ref mut param) => {
+                                    CapturedParam::Ident(param) => {
                                         obfuscate_ident(param);
                                     }
                                     _ => {}
@@ -47,7 +47,7 @@ fn obfuscate_params(params: &mut Punctuated<GenericParam, Comma>) {
                     }
                 }
             }
-            GenericParam::Const(ref mut param) => {
+            GenericParam::Const(param) => {
                 obfuscate_ident(&mut param.ident);
             }
         }